Refactor sources folder to group by module
