+++ services/gateway/apis/http-version3/methods/modules/legacy/legacyAccounts.js
@@ -13,7 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const legacyAccountsSource = require('../../../../../sources/version3/legacyAccountsSchema');
+const legacyAccountsSource = require('../../../../../sources/version3/modules/legacy/legacyAccountsSchema');
 const regex = require('../../../../../shared/regex');
 const { transformParams, response, getSwaggerDescription } = require('../../../../../shared/utils');
