+++ services/gateway/apis/http-version3/methods/modules/nft/nft.js
@@ -13,7 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const nftSource = require('../../../../../sources/version3/nft');
+const nftSource = require('../../../../../sources/version3/modules/nft/nft');
 const envelope = require('../../../../../sources/version3/mappings/stdEnvelope');
 const regex = require('../../../../../shared/regex');
 const { transformParams, response, getSwaggerDescription } = require('../../../../../shared/utils');
