+++ services/gateway/apis/http-version3/methods/modules/pos/posConstants.js
@@ -14,7 +14,7 @@
  *
  */
 
-const posConstantsSource = require('../../../../../sources/version3/posConstants');
+const posConstantsSource = require('../../../../../sources/version3/modules/pos/posConstants');
 const envelope = require('../../../../../sources/version3/mappings/stdEnvelope');
 const { response, getSwaggerDescription } = require('../../../../../shared/utils');
