+++ services/gateway/apis/http-version3/methods/modules/pos/posRewardsLocked.js
@@ -13,7 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const posRewardsLockedSource = require('../../../../../sources/version3/posRewardsLocked');
+const posRewardsLockedSource = require('../../../../../sources/version3/modules/pos/posRewardsLocked');
 const envelope = require('../../../../../sources/version3/mappings/stdEnvelope');
 const regex = require('../../../../../shared/regex');
 const { transformParams, response, getSwaggerDescription } = require('../../../../../shared/utils');
