+++ services/gateway/apis/http-version3/methods/modules/pos/posStakers.js
@@ -13,7 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const posStakersSource = require('../../../../../sources/version3/posStakers');
+const posStakersSource = require('../../../../../sources/version3/modules/pos/posStakers');
 const envelope = require('../../../../../sources/version3/mappings/stdEnvelope');
 const regex = require('../../../../../shared/regex');
 const { transformParams, response, getSwaggerDescription } = require('../../../../../shared/utils');
