+++ services/gateway/apis/http-version3/methods/modules/pos/posStakes.js
@@ -13,7 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const posStakesSource = require('../../../../../sources/version3/posStakes');
+const posStakesSource = require('../../../../../sources/version3/modules/pos/posStakes');
 const envelope = require('../../../../../sources/version3/mappings/stdEnvelope');
 const regex = require('../../../../../shared/regex');
 const { transformParams, response, getSwaggerDescription } = require('../../../../../shared/utils');
