+++ services/gateway/apis/http-version3/methods/modules/token/tokenBalances.js
@@ -13,7 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const tokenBalancesSource = require('../../../../../sources/version3/tokenBalances');
+const tokenBalancesSource = require('../../../../../sources/version3/modules/token/tokenBalances');
 const envelope = require('../../../../../sources/version3/mappings/stdEnvelope');
 const regex = require('../../../../../shared/regex');
 const { transformParams, response, getSwaggerDescription } = require('../../../../../shared/utils');
