+++ services/gateway/sources/version3/modules/auth/auth.js
@@ -1,6 +1,6 @@
 /*
  * LiskHQ/lisk-service
- * Copyright © 2023 Lisk Foundation
+ * Copyright © 2022 Lisk Foundation
  *
  * See the LICENSE file at the top-level directory of this distribution
  * for licensing information.
@@ -13,14 +13,20 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const nftSupported = require('./mappings/nftSupported');
+const auth = require('../../mappings/auth');
 
 module.exports = {
 	type: 'moleculer',
-	method: 'indexer.nft.supported',
-	params: {},
+	method: 'indexer.auth',
+	params: {
+		address: '=,string',
+	},
 	definition: {
-		data: nftSupported,
-		meta: {},
+		data: auth,
+		meta: {
+			address: '=,string',
+			publicKey: '=,string',
+			name: '=,string',
+		},
 	},
 };
