+++ services/gateway/sources/version3/modules/dynamicReward/rewardAnnualInflation.js
