+++ services/gateway/sources/version3/modules/dynamicReward/rewardConstants.js
