+++ services/gateway/sources/version3/modules/pos/posRewardsLocked.js
@@ -1,6 +1,6 @@
 /*
  * LiskHQ/lisk-service
- * Copyright © 2023 Lisk Foundation
+ * Copyright © 2022 Lisk Foundation
  *
  * See the LICENSE file at the top-level directory of this distribution
  * for licensing information.
@@ -13,14 +13,24 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const tokenConstants = require('./mappings/tokenConstants');
+const posRewardLocked = require('../../mappings/posRewardsLocked');
 
 module.exports = {
 	type: 'moleculer',
-	method: 'indexer.token.constants',
-	params: {},
+	method: 'indexer.pos.rewards.locked',
+	params: {
+		address: '=,string',
+		publicKey: '=,string',
+		name: '=,string',
+		limit: '=,number',
+		offset: '=,number',
+	},
 	definition: {
-		data: tokenConstants,
-		meta: {},
+		data: ['data', posRewardLocked],
+		meta: {
+			count: '=,number',
+			offset: '=,number',
+			total: '=,number',
+		},
 	},
 };
