+++ services/gateway/sources/version3/modules/token/tokenBalances.js
@@ -13,25 +13,24 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const {
-	escrowedAmount,
-	supportedToken,
-	totalSupplyByToken,
-} = require('./mappings/tokenSummary');
+const token = require('../../mappings/token');
 
 module.exports = {
 	type: 'moleculer',
-	method: 'indexer.token.summary',
+	method: 'indexer.token.balances',
 	params: {
+		address: '=,string',
+		tokenID: '=,string',
 		offset: '=,number',
 		limit: '=,number',
 	},
 	definition: {
-		data: {
-			escrowedAmounts: ['data.escrowedAmounts', escrowedAmount],
-			supportedTokens: supportedToken,
-			totalSupply: ['data.totalSupply', totalSupplyByToken],
+		data: ['data', token],
+		meta: {
+			address: '=,string',
+			count: '=,number',
+			offset: '=,number',
+			total: '=,number',
 		},
-		meta: {},
 	},
 };
