+++ services/gateway/sources/version3/modules/validator/validatorValidateBLSKey.js
